fix(equipment): expose overflow options as checkboxes
